Covering a failure during the housekeeping stage
Starting to add specifc tests for each state.
