+++ pocs/tests/conftest.py
@@ -8,8 +8,12 @@ import copy
 import os
 import pytest
 
-import pocs.base
+from pocs import base
+from pocs.core import POCS
 from pocs.utils.config import load_config
+from pocs.camera import create_cameras_from_config
+from pocs.observatory import Observatory
+
 
 # Global variable with the default config; we read it once, copy it each time it is needed.
 _one_time_config = None
@@ -17,7 +21,7 @@ _one_time_config = None
 
 @pytest.fixture(scope='function')
 def config():
-    pocs.base.reset_global_config()
+    base.reset_global_config()
 
     global _one_time_config
     if not _one_time_config:
@@ -41,3 +45,46 @@ def config_with_simulated_dome(config):
 @pytest.fixture
 def data_dir():
     return '{}/pocs/tests/data'.format(os.getenv('POCS'))
+
+
+@pytest.fixture(scope='function')
+def cameras(config):
+    """Get the default cameras from the config."""
+    return create_cameras_from_config(config)
+
+
+@pytest.fixture(scope='function')
+def observatory(config, db_type, cameras):
+    observatory = Observatory(
+        config=config,
+        cameras=cameras,
+        simulator=['all'],
+        ignore_local_config=True,
+        db_type=db_type
+    )
+    return observatory
+
+
+@pytest.fixture(scope='function')
+def pocs(config, observatory):
+    os.environ['POCSTIME'] = '2016-08-13 13:00:00'
+
+    p = POCS(observatory,
+             run_once=True,
+             config=config,
+             ignore_local_config=True)
+
+    p.observatory.scheduler.fields_file = None
+    p.observatory.scheduler.fields_list = [
+        {'name': 'Wasp 33',
+         'position': '02h26m51.0582s +37d33m01.733s',
+         'priority': '100',
+         'exp_time': 2,
+         'min_nexp': 2,
+         'exp_set_size': 2,
+         },
+    ]
+
+    yield p
+
+    p.power_down()
